PMM-2038 percona accessors move to separate file
